requirements.txt
greenlet==3.1.1
playwright==1.49.0
pyee==12.0.0
python-dotenv==1.0.1
tk==0.1.0
typing_extensions==4.12.2